fix: evalPoints 컬럼 INTEGER → NUMERIC(5,2) — 0.5점 카테고리 INSERT 실패 해결

+++ prisma/init.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS "patrol_entries" (
   "userId"            TEXT NOT NULL REFERENCES "users"("id"),
   "category"          TEXT NOT NULL,
   "evalItem"          TEXT,
-  "evalPoints"        INTEGER,
+  "evalPoints"        NUMERIC(5,2),
   "latitude"          DOUBLE PRECISION,
   "longitude"         DOUBLE PRECISION,
   "address"           TEXT,
